No need for get_dependencies() to recurse.
The recursion logic is responsibility of the Resolver class already. The
PackageManager can be kept simple and return only the direct
dependencies for a given (name, version) combination.

I also moved the caching of the result to the get_dependencies() method
directly, instead of caching a method called extract_dependencies().