Store (name, version)->[deps] mappings in depcache.
WARNING: Please do remove your ~/.pip-tools folder before using this!

This change really speeds up things.  Previously, in find_best_match(),
we had to reach out to PyPI for all specs, even for pinned ones.  In
those cases, there were really only two outcomes: either the pinned
version did exist (and we get a link), or it did not (in which case
there's a package not found exception due).

As reaching out to PyPI seemed like a pretty heavy overhead for such
a simple query, this patch changes the dependency cache to store (name,
version) tuples as keys (instead of local paths), which a) makes more
sense anyway, and b) allows the find_best_match() lookup to "cheat" for
pinned versions: if it's in the dep cache, the package must exist and
was previously downloaded, unpacked and inspected.

In other words: if we ask find_best_match('Django', '1.4.1') and that's
in the cache, we can immediately return '1.4.1'.