Support the != operator.
It is a very rare one, but it occurs in the wild, for example, as
a nested dependency from Fabric==1.4.3 (a popular package):

pycrypto>=2.1,!=2.4 (from requirements.in:9 ~> Fabric==1.4.3 ~> ssh==1.7.14)

This complicates the already complicated normalization code, and
although I've added unit tests around this, I strongly believe that with
the right abstraction, this code can be hugely simplified.  Ideas are
welcome.