Upgrade to container builds on travis-ci
Per the
[docs](https://docs.travis-ci.com/user/migrating-from-legacy/#Why-migrate-to-container-based-infrastructure%3F),
container based builds are faster and provide caching. This should (in
theory) allow for faster builds.

Builds are currently very quick (8-15 seconds per environment), but they
seem to take a significant amount of time to begin (spent waiting in the
build queue). I don't see any documentation on whether or not container
based builds have an effect on time spent in the build queue, so this is
an experiment of sorts. ``sudo`` is not currently used in the travis
builds, so there are no downsides to upgrading to the new container
infrastructure.