🧪 Add "supported" pip in tox for PR/push @ CI
This is meant to be separate from "latest" and is pinned to the known
working released version of pip. It will run on merges to `main` and
in pull requests. And the cron runs will additionally test the latest
version on PyPI and the `main` branch of the pip repo.