Restrict dependencies; improve CI reproducibility
`click` v8.2 needs to be handled -- as a follow-up change -- but it's
hard to know if the application itself will work with it yet. Instead
of restricting that in tox.ini, I have chosen to restrict it in the
metadata itself.

`setuptools` on newer versions breaks some of the tests, but
restricting it for package metadata is definitely wrong. Therefore, it
is only restricted for now on the `pipsupported` build.

This should lead to improved CI, but will not result in all tests
passing.