Fix compile on `-r` relative paths outside of cwd
When `-r` is used to pass a path, we are normalizing with
`pathlib.Path.relative_to`. This fails when the input is not a subpath
of the current working directory.

In Python 3.12+ pathlib supports this usage (`walk_up=True`), but on
older Pythons we need to fallback to using `os.path.relpath`.

In order to support this usage and give a clear indication of how we
should upgrade when older versions are dropped, a small compat wrapper
is here defined, `_relative_to_walk_up` which uses
`relative_to(..., walk_up=True)` when it is available, and only uses
`os.path.relpath` when necessary. After we drop Python 3.11 in a few
years, the helper can be removed.

A new regression test reproduces #2231 and is fixed with this patch.
As part of writing that test, I wanted to leverage the
`TestFilesCollection` helper, so this was extended in a small way to
support usage without an explicit name.