Add support for builds from 'git archive'
setuptools-scm needs the git metadata expanded into this file in order to
support installation from an archive. Providing the data makes it
possible for `setuptools-scm` to build cleanly from an archive.

setuptools-scm docs recommend using the manifest file to explicitly skip
`.git_archival.txt` when building sdists, as its inclusion results in
build-time warnings during development.

The pattern used in `.git_archival.txt` as a tag match has to be
something expressible in glob syntax. Given this restriction, it simply
requires a prefix of `v\d` and a suffix of `\d` (in appropriate syntax)
with any midfix allowed.